+++ src/Makefile.in
@@ -133,7 +133,7 @@ SRCS = \
 	common/io/limitRead.c \
 	common/io/server.c \
 	common/io/session.c \
-	common/io/socket/addressInfo.c \
+	common/io/socket/address.c \
 	common/io/socket/client.c \
 	common/io/socket/common.c \
 	common/io/socket/server.c \
