+++ src/common/io/socket/address.c
@@ -4,9 +4,9 @@ Address Lookup
 #include "build.auto.h"
 
 #include "common/debug.h"
+#include "common/io/socket/address.h"
 #include "common/log.h"
 #include "common/memContext.h"
-#include "common/io/socket/addressInfo.h"
 
 /***********************************************************************************************************************************
 Object type
