+++ test/define.yaml
@@ -380,7 +380,7 @@ unit:
           - common/io/tls/common
           - common/io/tls/server
           - common/io/tls/session
-          - common/io/socket/addressInfo
+          - common/io/socket/address
           - common/io/socket/client
           - common/io/socket/common
           - common/io/socket/server
